requirements.txt
requests
backoff
pandas
asyncio
aiohttp
python-dotenv
sqlalchemy
pytest
pytest-vcr
psycopg2-binary